+++ lib/ex_nylas/api_keys/api_keys.ex
@@ -10,6 +10,7 @@ defmodule ExNylas.APIKeys do
     APIKey,
     Connection,
     DecodeError,
+    Error,
     ErrorHandler,
     Response,
     ResponseHandler,
@@ -25,7 +26,7 @@ defmodule ExNylas.APIKeys do
       iex> {:ok, result} = ExNylas.APIKeys.create(conn, "application_id", %{name: "My API Key", expires_in: 90  }, "signature", "kid", "nonce", "timestamp")
   """
   @spec create(Connection.t(), String.t(), map(), String.t(), String.t(), String.t(), String.t()) ::
-    {:ok, Response.t()} | {:error, Response.t() | TransportError.t() | DecodeError.t()}
+    {:ok, Response.t()} | {:error, Response.t() | TransportError.t() | DecodeError.t() | Error.t()}
   def create(%Connection{} = conn, application_id, body, signature, kid, nonce, timestamp) do
     Req.new(
       url: "#{conn.api_server}/v3/admin/applications/#{application_id}/api-keys",
@@ -61,7 +62,7 @@ defmodule ExNylas.APIKeys do
       iex> {:ok, result} = ExNylas.APIKeys.list(conn, "application_id", "signature", "kid", "nonce", "timestamp")
   """
   @spec list(Connection.t(), String.t(), String.t(), String.t(), String.t(), String.t()) ::
-    {:ok, Response.t()} | {:error, Response.t() | TransportError.t() | DecodeError.t()}
+    {:ok, Response.t()} | {:error, Response.t() | TransportError.t() | DecodeError.t() | Error.t()}
   def list(%Connection{} = conn, application_id, signature, kid, nonce, timestamp) do
     Req.new(
       url: "#{conn.api_server}/v3/admin/applications/#{application_id}/api-keys",
@@ -96,7 +97,7 @@ defmodule ExNylas.APIKeys do
       iex> {:ok, result} = ExNylas.APIKeys.find(conn, "application_id", "api_key_id", "signature", "kid", "nonce", "timestamp")
   """
   @spec find(Connection.t(), String.t(), String.t(), String.t(), String.t(), String.t(), String.t()) ::
-    {:ok, Response.t()} | {:error, Response.t() | TransportError.t() | DecodeError.t()}
+    {:ok, Response.t()} | {:error, Response.t() | TransportError.t() | DecodeError.t() | Error.t()}
   def find(%Connection{} = conn, application_id, api_key_id, signature, kid, nonce, timestamp) do
     Req.new(
       url: "#{conn.api_server}/v3/admin/applications/#{application_id}/api-keys/#{api_key_id}",
@@ -131,7 +132,7 @@ defmodule ExNylas.APIKeys do
       iex> {:ok, result} = ExNylas.APIKeys.delete(conn, "application_id", "api_key_id", "signature", "kid", "nonce", "timestamp")
   """
   @spec delete(Connection.t(), String.t(), String.t(), String.t(), String.t(), String.t(), String.t()) ::
-    {:ok, Response.t()} | {:error, Response.t() | TransportError.t() | DecodeError.t()}
+    {:ok, Response.t()} | {:error, Response.t() | TransportError.t() | DecodeError.t() | Error.t()}
   def delete(%Connection{} = conn, application_id, api_key_id, signature, kid, nonce, timestamp) do
     Req.new(
       url: "#{conn.api_server}/v3/admin/applications/#{application_id}/api-keys/#{api_key_id}",