+++ lib/ex_nylas/core/paging/cursor.ex
@@ -4,6 +4,7 @@ defmodule ExNylas.Paging.Cursor do
   alias ExNylas.{
     Connection,
     DecodeError,
+    Error,
     Paging.Helpers,
     Paging.Options,
     Response,
@@ -14,11 +15,11 @@ defmodule ExNylas.Paging.Cursor do
           Connection.t(),
           (Connection.t(), Keyword.t() | map() ->
              {:ok, Response.t()}
-             | {:error, Response.t() | TransportError.t() | DecodeError.t()}),
+             | {:error, Response.t() | TransportError.t() | DecodeError.t() | Error.t()}),
           Keyword.t() | map()
         ) ::
           {:ok, [struct()]}
-          | {:error, Response.t() | TransportError.t() | DecodeError.t()}
+          | {:error, Response.t() | TransportError.t() | DecodeError.t() | Error.t()}
   def all(conn, list_function, opts \\ []) do
     page_with_cursor(conn, list_function, Options.from_opts(opts))
   end
@@ -27,13 +28,13 @@ defmodule ExNylas.Paging.Cursor do
           Connection.t(),
           (Connection.t(), Keyword.t() | map() ->
              {:ok, Response.t()}
-             | {:error, Response.t() | TransportError.t() | DecodeError.t()}),
+             | {:error, Response.t() | TransportError.t() | DecodeError.t() | Error.t()}),
           Keyword.t() | map(),
           any() | nil,
           [any()]
         ) ::
           {:ok, [any()]}
-          | {:error, Response.t() | TransportError.t() | DecodeError.t()}
+          | {:error, Response.t() | TransportError.t() | DecodeError.t() | Error.t()}
   defp page_with_cursor(%Connection{} = conn, list_function, opts, next_cursor \\ nil, acc \\ []) do
     %{query: query, delay: delay, send_to: send_to, with_metadata: with_metadata} = opts
     query = put_in(query, [:page_token], next_cursor)