+++ lib/ex_nylas/core/paging/paging.ex
@@ -4,6 +4,7 @@ defmodule ExNylas.Paging do
   alias ExNylas.{
     Connection,
     DecodeError,
+    Error,
     Paging.Cursor,
     Paging.Offset,
     Response,
@@ -14,12 +15,12 @@ defmodule ExNylas.Paging do
           Connection.t(),
           (Connection.t(), Keyword.t() | map() ->
              {:ok, Response.t()}
-             | {:error, Response.t() | TransportError.t() | DecodeError.t()}),
+             | {:error, Response.t() | TransportError.t() | DecodeError.t() | Error.t()}),
           boolean(),
           Keyword.t() | map()
         ) ::
           {:ok, [struct()]}
-          | {:error, Response.t() | TransportError.t() | DecodeError.t()}
+          | {:error, Response.t() | TransportError.t() | DecodeError.t() | Error.t()}
   def all(conn, list_function, use_cursor_paging, opts \\ [])
   def all(conn, list_function, true = _use_cursor_paging, opts) do
     Cursor.all(conn, list_function, opts)
@@ -33,7 +34,7 @@ defmodule ExNylas.Paging do
           Connection.t(),
           (Connection.t(), Keyword.t() | map() ->
              {:ok, Response.t()}
-             | {:error, Response.t() | TransportError.t() | DecodeError.t()}),
+             | {:error, Response.t() | TransportError.t() | DecodeError.t() | Error.t()}),
           boolean(),
           Keyword.t() | map()
         ) :: [struct()]